requirements.txt
black==22.1.0
certifi==2021.10.8
confluent-kafka[avro,json,protobuf]>=1.4.2
click==8.0.3
PyYAML==6.0
requests==2.27.1
